+++ CHANGELOG.md
@@ -25,6 +25,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Fixed an unnecessary CAPMC API request and a confusing warning message during
+  `sat bootsys shutdown --stage cabinet-power` when there are no non-management
+  nodes in air-cooled cabinets.
+
 ## [3.22.0] - 2023-05-08
 
 ### Added